feat: added CLI Error Handler & moved logger to tools dir
