+++ apps/cli/src/lib/util/index.ts
@@ -1,6 +1,5 @@
 export * from './ApiService';
 export * from './BrowserService';
-export * from './CliLoggerService';
 export * from './GlobalConfigService';
 export * from './Helpers';
 export * from './ProjectConfigService';
