+++ apps/cli/src/main.ts
@@ -1,8 +1,12 @@
 import { CommandFactory } from 'nest-commander';
 import { AppModule } from './modules/app/app.module';
+import { handleException } from './tools';
 
 const bootstrap = async () => {
-  await CommandFactory.run(AppModule);
+  await CommandFactory.run(AppModule, {
+    cliName: 'The Secret Store',
+    errorHandler: handleException,
+  });
 };
 
 bootstrap();
