+++ apps/cli/src/tools/index.ts
@@ -1,0 +1,2 @@
+export * from './CliLoggerService';
+export * from './ErrorHandler';
